refactor(ssh-keys): rename sshKeys folders
