+++ solutions/swb-app/config/jest.config.json
@@ -16,7 +16,7 @@
     "<rootDir>/src/environmentTypeRoutes.ts",
     "<rootDir>/src/generateRouter.ts",
     "<rootDir>/src/index.ts",
-    "<rootDir>/src/keyPairRoutes.ts",
+    "<rootDir>/src/sshKeyRoutes.ts",
     "<rootDir>/src/projectRoutes.ts",
     "<rootDir>/src/staticPermissionsConfig.ts",
     "<rootDir>/src/staticRouteConfig.ts",
@@ -45,15 +45,15 @@
     "<rootDir>/src/projectEnvTypeConfigs/projectEnvTypeConfigPlugin.ts",
     "<rootDir>/src/projectEnvTypeConfigs/projectEnvTypeConfigRoutes.ts",
 
-    "<rootDir>/src/keyPairs/createKeyPairRequest.ts",
-    "<rootDir>/src/keyPairs/createKeyPairResponse.ts",
-    "<rootDir>/src/keyPairs/deleteKeyPairRequest.ts",
-    "<rootDir>/src/keyPairs/getKeyPairRequest.ts",
-    "<rootDir>/src/keyPairs/getKeyPairResponse.ts",
-    "<rootDir>/src/keyPairs/keyPair.ts",
-    "<rootDir>/src/keyPairs/keyPairPlugin.ts",
-    "<rootDir>/src/keyPairs/sendPublicKeyRequest.ts",
-    "<rootDir>/src/keyPairs/sendPublicKeyResponse.ts",
+    "<rootDir>/src/sshKeys/createSshKeyRequest.ts",
+    "<rootDir>/src/sshKeys/createSshKeyResponse.ts",
+    "<rootDir>/src/sshKeys/deleteSshKeyRequest.ts",
+    "<rootDir>/src/sshKeys/listUserSshKeysForProjectRequest.ts",
+    "<rootDir>/src/sshKeys/listUserSshKeysForProjectResponse.ts",
+    "<rootDir>/src/sshKeys/sshKey.ts",
+    "<rootDir>/src/sshKeys/sshKeyPlugin.ts",
+    "<rootDir>/src/sshKeys/sendPublicKeyRequest.ts",
+    "<rootDir>/src/sshKeys/sendPublicKeyResponse.ts",
 
     "<rootDir>/src/validatorHelper.ts",
     "<rootDir>/src/errorHandlers.ts"
